Delete redundant interface, fixed inconsistent naming etc. cosmetics (addressing further remarks from code review)
